+++ yash-builtin/CHANGELOG.md
@@ -20,8 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- The `eval`, `exit`, `return`, and `shift` built-ins now correctly handle the
-  `--` separator between options and operands.
+- The `eval`, `exit`, `return`, `shift`, and `typeset` built-ins now correctly
+  handle the `--` separator between options and operands.
 - The `jobs` built-in no longer panics when reporting the same finished job more
   than once in a single invocation.