+++ yash-builtin/CHANGELOG.md
@@ -20,8 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- The `eval`, `exit`, `return`, `shift`, and `typeset` built-ins now correctly
-  handle the `--` separator between options and operands.
+- The `eval`, `exit`, `export`, `readonly`, `return`, `shift`, and `typeset`
+  built-ins now correctly handle the `--` separator between options and
+  operands.
 - The `export`, `readonly`, and `typeset` built-ins now correctly print the `--`
   separator when the name of a variable or function starts with `-`.
 - The `jobs` built-in no longer panics when reporting the same finished job more