+++ yash-builtin/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - External dependency versions:
+    - yash-semantics (optional) 0.8.0 → 0.8.1
     - yash-syntax 0.15.0 → 0.15.1
 
 ### Fixed
@@ -40,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - External dependency versions:
     - yash-env 0.8.0 → 0.9.0
-    - yash-semantics (optional) 0.8.0 → 0.9.0
+    - yash-semantics (optional) 0.7.1 → 0.8.0
     - yash-syntax 0.14.1 → 0.15.0
 - Internal dependency versions:
     - yash-prompt (optional) 0.5.0 → 0.6.0