+++ yash-builtin/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to `yash-builtin` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - Unreleased
+
+- External dependency versions:
+    - yash-semantics (optional) 0.8.1 → 0.9.0
+
 ## [0.9.1] - 2025-09-20
 
 ### Added
@@ -358,6 +363,7 @@ The `wait` built-in no longer treats suspended jobs as terminated jobs.
 
 - Initial implementation of the `yash-builtin` crate
 
+[0.10.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.10.0
 [0.9.1]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.9.1
 [0.9.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.9.0
 [0.8.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.8.0